converted styling in login.js into semantic ui react, renamed the login to Login.jsx
